refactor: partie planning complète
