--- src/features/Planning/views/WeekView.tsx
@@ -1,292 +1,0 @@
-import React from 'react';
-import { FlatList, Text, TouchableOpacity, View } from 'react-native';
-import { MaterialCommunityIcons, MaterialIcons } from '@expo/vector-icons';
-import moment from 'moment';
-import { SafeAreaView } from 'react-native-safe-area-context';
-
-import CalendarWeek from '../components/CalendarWeek';
-import { WeekComponent } from '../components/ScheduleList'; 
-import style, { tokens } from '../../../shared/theme/Theme'; 
-import Translator from '../../../shared/i18n/Translator';
-import { AppContext } from '../../../shared/services/AppCore';
-
-const isEqualsObject = (obj1, ojb2) => {
-	// objects must be in the same order
-	return Object.entries(obj1).toString() === Object.entries(ojb2).toString();
-};
-
-const findIndexOfObject = (objectList, object) => {
-	for (let i = 0; i < objectList.length; i++) {
-		const element = objectList[i];
-		if (isEqualsObject(element, object)) {
-			return i;
-		}
-	}
-	return -1;
-};
-
-export interface WeekViewProps {
-    navigation: import('@react-navigation/native').NavigationProp<Record<string, unknown>> & { setOptions: (options: unknown) => void };
-    route: { params: Record<string, unknown> };
-    groupName?: string | string[];
-}
-
-export interface WeekViewState {
-    currentWeek: { week: number; year: number };
-    currentWeekIndex: number;
-    weeks: { week: number; year: number }[];
-    selectedWeek: { week: number; year: number };
-}
-
-class WeekView extends React.Component<WeekViewProps, WeekViewState> {
-	static contextType = AppContext;
-	declare context: React.ContextType<typeof AppContext>;
-    viewability: import('react-native').ViewToken[];
-    calendarList: import('react-native').FlatList | null;
-    _lastTitle: string | null = null;
-
-	constructor(props: WeekViewProps) {
-		super(props);
-
-		const currentDate = moment();
-		const currentWeek = { week: currentDate.isoWeek(), year: currentDate.year() };
-		const weeks = WeekView.generateWeeks();
-
-		this.state = {
-			currentWeek: currentWeek,
-			currentWeekIndex: findIndexOfObject(weeks, currentWeek),
-			weeks,
-			selectedWeek: currentWeek,
-		};
-
-		this.viewability = {
-			itemVisiblePercentThreshold: 50,
-		};
-	}
-
-	getGroupName() {
-		let groupName = this.props.route.params.groupName;
-		if (Array.isArray(groupName)) {
-			groupName = this.context.favoriteGroups;
-		}
-		return groupName;
-	}
-
-	componentDidMount() {
-		this._lastTitle = null;
-		this.updateTitle();
-	}
-
-	componentDidUpdate() {
-		this.updateTitle();
-	}
-
-	updateTitle() {
-		const { treatTitle } = require('../../../shared/services/AppCore');
-		const title = treatTitle(this.getGroupName());
-		if (title !== this._lastTitle) {
-			this._lastTitle = title;
-			this.props.navigation.setOptions({ title });
-		}
-	}
-
-	static getCalendarListItemLayout = (data, index) => {
-		return {
-			length: style.calendarList.itemSize,
-			offset: style.calendarList.itemSize * index,
-			index,
-		};
-	};
-
-	renderCalendarListItem = ({ item }) => {
-		return (
-			<CalendarWeek
-				week={item}
-				selectedWeek={this.state.selectedWeek}
-				currentWeek={this.state.currentWeek}
-				onPressItem={this.onWeekPress}
-				theme={style.Theme[this.context.themeName]}
-			/>
-		);
-	};
-
-	extractCalendarListItemKey = (item) => {
-		return `S${item.week}-${this.context.themeName}`;
-	};
-
-	onTodayPress = () => {
-		this.setState(
-			{
-				selectedWeek: this.state.currentWeek,
-			},
-			() => {
-				if (this.calendarList) {
-					this.calendarList.scrollToIndex({
-						index: this.state.currentWeekIndex,
-						animated: true,
-					});
-				}
-			},
-		);
-	};
-
-	onDayButton = () => {
-		this.props.navigation.goBack();
-	};
-
-	onWeekPress = (item) => {
-		this.setState({
-			selectedWeek: item,
-		});
-	};
-
-	static generateWeeks = () => {
-		const currentDate = moment();
-		const beginningGenerationDate = moment()
-			.date(1)
-			.month(7);
-
-		if (currentDate.month() > 6) {
-			beginningGenerationDate.year(currentDate.year());
-		} else {
-			beginningGenerationDate.year(currentDate.year() - 1);
-		}
-
-		const weeks = [];
-		let firstWeek = null;
-
-		for (let i = 0, iMax = 365; i < iMax; i += 7) {
-			const date = moment(beginningGenerationDate).add(i, 'd');
-			const week = date.isoWeek();
-			const year = date.year();
-			if (week !== firstWeek) {
-				if (firstWeek === null) {
-					firstWeek = week;
-				}
-				weeks.push({ week, year });
-			} else {
-				break;
-			}
-		}
-
-		return weeks;
-	};
-
-	render() {
-		const theme = style.Theme[this.context.themeName];
-
-		return (
-            <SafeAreaView edges={['left', 'right']} style={{ flex: 1, backgroundColor: theme.courseBackground }}>
-                <WeekComponent
-                    key={`weekComponent-${this.context.themeName}`}
-                    week={this.state.selectedWeek}
-                    groupName={this.getGroupName()}
-                    theme={theme}
-                    navigation={this.props.navigation}
-                    filtersList={this.context.filters}
-                />
-
-                <SafeAreaView edges={['bottom']} style={{
-                    flexGrow: 0,
-                    backgroundColor: theme.cardBackground,
-                    borderTopWidth: 1,
-                    borderTopColor: theme.border,
-                }}>
-                    {/* Header : label + boutons */}
-                    <View style={{
-                        flexDirection: 'row',
-                        justifyContent: 'space-between',
-                        alignItems: 'center',
-                        height: 44,
-                        backgroundColor: theme.cardBackground,
-                        paddingHorizontal: tokens.space.xs + 2,
-                    }}>
-                        {/* Bouton Cette semaine */}
-                        <TouchableOpacity
-                            onPress={this.onTodayPress}
-                            style={{
-                                flexDirection: 'row',
-                                alignItems: 'center',
-                                paddingHorizontal: tokens.space.md,
-                                paddingVertical: tokens.space.sm,
-                                borderRadius: tokens.radius.md,
-                                backgroundColor: theme.greyBackground,
-                            }}>
-                            <MaterialIcons
-                                name="event-note"
-                                size={16}
-                                color={theme.accent ?? theme.primary}
-                            />
-                            <Text style={{
-                                fontSize: tokens.fontSize.sm,
-                                marginLeft: tokens.space.xs,
-                                color: theme.accent ?? theme.primary,
-                                fontWeight: tokens.fontWeight.medium,
-                            }}>
-                                {Translator.get('THIS_WEEK')}
-                            </Text>
-                        </TouchableOpacity>
-
-                        {/* Label semaine */}
-                        <Text style={{
-                            fontSize: tokens.fontSize.md,
-                            fontWeight: tokens.fontWeight.semibold,
-                            color: theme.fontSecondary,
-                            position: 'absolute',
-                            left: 0,
-                            right: 0,
-                            textAlign: 'center',
-                        }}>
-                            {Translator.get('WEEK')}
-                        </Text>
-
-                        {/* Bouton Jour */}
-                        <TouchableOpacity
-                            onPress={this.onDayButton}
-                            style={{
-                                flexDirection: 'row',
-                                alignItems: 'center',
-                                paddingHorizontal: tokens.space.md,
-                                paddingVertical: tokens.space.sm,
-                                borderRadius: tokens.radius.md,
-                                backgroundColor: theme.greyBackground,
-                            }}>
-                            <Text style={{
-                                fontSize: tokens.fontSize.sm,
-                                marginRight: tokens.space.xs,
-                                color: theme.accent ?? theme.primary,
-                                fontWeight: tokens.fontWeight.medium,
-                            }}>
-                                {Translator.get('DAY')}
-                            </Text>
-                            <MaterialCommunityIcons
-                                name="calendar"
-                                size={16}
-                                color={theme.accent ?? theme.primary}
-                            />
-                        </TouchableOpacity>
-                    </View>
-
-                    {/* Liste des semaines */}
-                    <FlatList
-                        ref={(list) => { this.calendarList = list; }}
-                        showsHorizontalScrollIndicator={false}
-                        data={this.state.weeks}
-                        horizontal={true}
-                        keyExtractor={this.extractCalendarListItemKey}
-                        viewabilityConfig={this.viewability}
-                        initialScrollIndex={this.state.currentWeekIndex}
-                        getItemLayout={WeekView.getCalendarListItemLayout}
-                        extraData={this.state}
-                        renderItem={this.renderCalendarListItem}
-                        style={{
-                            backgroundColor: theme.cardBackground,
-                        }}
-                    />
-                </SafeAreaView>
-            </SafeAreaView>
-        );
-	}
-}
-
-export default WeekView;
